stop_flex/requirements.txt
google-cloud-bigquery-reservation==0.2.0
google-api-core==1.19.0
Flask==1.1.2